+++ CHANGES.md
@@ -57,6 +57,7 @@ New:
 - Added syntax for `for` and `while` loops (#1252).
 - Added a bunch of source-related methods (#1379).
 - Added `min` and `max` functions.
+- Added `lufs` to compute the LUFS loundness (#1497).
 
 Changed: